Workout progress WIP

+++ app/src/main/java/ru/ok/technopolis/training/personal/items/BundleItem.kt
@@ -1,7 +1,7 @@
 package ru.ok.technopolis.training.personal.items
 
 data class BundleItem(
-    override val id: String,
-    val itemId: Int,
-    val title: String
+        override val id: String,
+        val itemId: Long,
+        val title: String
 ) : SelectableItem(id)
